+++ R/doc-helper-cli.R
@@ -25,19 +25,19 @@
 #' The model returns approximately the same number of output tokens as it
 #' receives, so we'll call that 100 output tokens per refactor.
 #'
-#' As of the time of writing (October 2024), the recommended chores model Claude
-#' Sonnet 3.5 costs $3 per million input tokens and $15 per million output
+#' As of the time of writing, the recommended chores model Claude
+#' Haiku 4.5 costs $1 per million input tokens and $5 per million output
 #' tokens. So, using the recommended model,
-#' **cli helpers cost around $9 for every 1,000 refactored pieces of code**. GPT-4o
+#' **cli helpers cost around $3 for every 1,000 refactored pieces of code**. GPT-4.1
 #' Mini, by contrast, doesn't tend to get cli markup classes right but _does_
 #' return syntactically valid calls to cli functions, and it would cost around
-#' 45 cents per 1,000 refactored pieces of code.
+#' $1 per 1,000 refactored pieces of code.
 #'
 #' @section Gallery:
 #'
 #' This section includes a handful of examples
 #' ["from the wild"](https://github.com/tidymodels/tune/blob/f8d734ac0fa981fae3a87ed2871a46e9c40d509d/R/checks.R)
-#' and are generated with the recommended model, Claude Sonnet 3.5.
+#' and are generated with Claude Sonnet 3.5.
 #'
 #' At its simplest, a one-line message with a little bit of markup:
 #'