+++ R/init-helper.R
@@ -13,7 +13,7 @@
 #' @param .chores_chat An ellmer Chat, e.g.
 #' `ellmer::chat_claude()`. Defaults to the `chores.chat` option,
 #' so e.g. set
-#' `options(chores.chat = ellmer::chat_claude(model = "claude-3-7-sonnet-20250219"))`
+#' `options(chores.chat = ellmer::chat_claude(model = "claude-haiku-4-5"))`
 #'  in your
 #' `.Rprofile` to configure chores with ellmer every time you start a new R session.
 #'
@@ -24,16 +24,16 @@
 #' # requires an API key and sets options
 #' \dontrun{
 #' # to create a chat with claude:
-#' .init_helper(.chores_chat = ellmer::chat_claude(model = "claude-3-7-sonnet-20250219"))
+#' .init_helper(.chores_chat = ellmer::chat_claude(model = "claude-haiku-4-5"))
 #'
-#' # or with OpenAI's GPT-4o-mini:
-#' .init_helper(.chores_chat = ellmer::chat_openai(model = "gpt-4o-mini"))
+#' # or with OpenAI's GPT-4.1-mini:
+#' .init_helper(.chores_chat = ellmer::chat_openai(model = "gpt-4.1-mini"))
 #'
-#' # to set OpenAI's GPT-4o-mini as the default model powering chores, for example,
+#' # to set OpenAI's GPT-4.1-mini as the default model powering chores, for example,
 #' # set the following option (possibly in your .Rprofile, if you'd like
 #' # them to persist across sessions):
 #' options(
-#'   chores.chat = ellmer::chat_openai(model = "gpt-4o-mini")
+#'   chores.chat = ellmer::chat_openai(model = "gpt-4.1-mini")
 #' )
 #' }
 #' @export