+++ tests/testthat/test-helper-add-remove.R
@@ -2,7 +2,7 @@ test_that("helper addition and removal works", {
   skip_if(identical(Sys.getenv("ANTHROPIC_API_KEY"), ""))
   skip_if_not_installed("withr")
   withr::local_options(
-    .chores_chat = ellmer::chat_claude(model = "claude-3-7-sonnet-20250219")
+    .chores_chat = ellmer::chat_claude(model = "claude-haiku-4-5")
   )
 
   boop_prompt <- "just reply with beep bop boop regardless of input"