+++ NEWS.md
@@ -1,3 +1,5 @@
+# chores (development version)
+
 # chores 0.3.0
 
 * Adjusted the default recommended models in documentation (#98).