+++ NEWS.md
@@ -1,5 +1,9 @@
 # chores (development version)
 
+* Adjusted the recommended models in documentation.
+
+    - **Anthropic**: Claude Sonnet 3.7 -> Claude Haiku 4.5.
+
 * `prompt_new()` gains an `overwrite` argument. When `overwrite = TRUE`,
   an existing prompt for the given chore will be overwritten with the provided `contents` (#103).