+++ NEWS.md
@@ -1,4 +1,4 @@
-# chores (development version)
+# chores 0.3.1
 
 * Adjusted the recommended models in documentation.